Switching to CompileSupport instead of ActionControllerMixin
