+++ lib/prawnto/template_handlers/compile_support.rb
@@ -1,8 +1,7 @@
 module Prawnto
-  module TemplateHandler
+  module TemplateHandlers
 
     class CompileSupport
-
       attr_reader :options
 
       def initialize(controller)
@@ -16,21 +15,28 @@ module Prawnto
       end
 
       def set_headers
-        set_pragma
-        set_cache_control
-        set_content_type
-        set_disposition
+        # unless defined?(ActionMailer) && defined?(ActionMailer::Base) && self.is_a?(ActionMailer::Base)
+          set_pragma
+          set_cache_control
+          set_content_type
+          set_disposition
+          set_other_headers_for_ie_ssl
+        # end
       end
 
-      # TODO: kept around from railspdf-- maybe not needed anymore? should check.
       def ie_request?
         @controller.request.env['HTTP_USER_AGENT'] =~ /msie/i
       end
 
-      # added to make ie happy with ssl pdf's (per naisayer)
       def ssl_request?
-        protocol = @controller.request.env['SERVER_PROTOCOL']
-        protocol && protocol.downcase == "https"
+        @controller.request.ssl?
+      end
+      
+      def set_other_headers_for_ie_ssl
+        return unless ssl_request? && ie_request?
+        headers['Content-Description'] = 'File Transfer'
+        headers['Content-Transfer-Encoding'] = 'binary'
+        headers['Expires'] = '0'        
       end
 
       # TODO: kept around from railspdf-- maybe not needed anymore? should check.
@@ -57,7 +63,7 @@ module Prawnto
 
       def set_disposition
         inline = options[:inline] ? 'inline' : 'attachment'
-        filename = options[:filename] ? "filename=#{options[:filename]}" : nil
+        filename = options[:filename] ? "filename=\"#{options[:filename]}\"" : nil
         @controller.headers["Content-Disposition"] = [inline,filename].compact.join(';')
       end
